Moved daily challenge settings into UserSettings.
